README.md
The portfolio website is a captivating online exhibit of an individual's or professional's achievements, skills, and creative ventures. Developed using HTML, CSS, and JavaScript, it delivers an engaging and interactive user experience. The design is responsive, adapting effortlessly to small screens, guaranteeing accessibility and user-friendliness. The concise yet comprehensive presentation allows visitors to easily explore and appreciate the creator's talents and capabilities, making it an indispensable asset for networking, job hunting, or personal branding.